feat[#361] : getProfile 서버 변경 사항에 맞춰 query 문 변경
