+++ src/entities/profile/api/index.ts
@@ -1,2 +1,1 @@
-// export * from "./getProfile";
-export * from "./_getProfile";
+export * from "./getProfile";
